backend/requirements.txt
# Pinned lower bounds support Python 3.14 (prebuilt wheels on Windows).
fastapi>=0.115.6
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
anthropic>=0.28.0
firebase-admin>=6.5.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
pydantic>=2.12.0
